feat: consolidate graphile-plugin-connection-filter-postgis into graphile-postgis
- Move createPostgisOperatorFactory() into graphile-postgis/src/plugins/connection-filter-operators.ts
- Move operator and integration tests into graphile-postgis/__tests__/
- Update GraphilePostgisPreset to include connectionFilterOperatorFactories
- Export createPostgisOperatorFactory from graphile-postgis
- Add graphile-connection-filter as optional peer dependency
- Update ConstructivePreset to import from graphile-postgis
- Remove graphile-plugin-connection-filter-postgis package entirely
- Remove graphile-settings dependency on old package
