--- graphile/graphile-plugin-connection-filter-postgis/CHANGELOG.md
@@ -1,66 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## [2.5.2](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter-postgis@2.5.1...graphile-plugin-connection-filter-postgis@2.5.2) (2026-03-12)
-
-**Note:** Version bump only for package graphile-plugin-connection-filter-postgis
-
-## [2.5.1](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter-postgis@2.4.0...graphile-plugin-connection-filter-postgis@2.5.1) (2026-03-12)
-
-**Note:** Version bump only for package graphile-plugin-connection-filter-postgis
-
-# [2.5.0](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter-postgis@2.4.0...graphile-plugin-connection-filter-postgis@2.5.0) (2026-03-12)
-
-**Note:** Version bump only for package graphile-plugin-connection-filter-postgis
-
-# [2.4.0](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter-postgis@2.3.3...graphile-plugin-connection-filter-postgis@2.4.0) (2026-03-12)
-
-**Note:** Version bump only for package graphile-plugin-connection-filter-postgis
-
-## [2.3.3](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter-postgis@2.3.2...graphile-plugin-connection-filter-postgis@2.3.3) (2026-03-12)
-
-**Note:** Version bump only for package graphile-plugin-connection-filter-postgis
-
-## [2.3.2](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter-postgis@2.3.1...graphile-plugin-connection-filter-postgis@2.3.2) (2026-03-04)
-
-**Note:** Version bump only for package graphile-plugin-connection-filter-postgis
-
-## [2.3.1](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter-postgis@2.3.0...graphile-plugin-connection-filter-postgis@2.3.1) (2026-03-03)
-
-**Note:** Version bump only for package graphile-plugin-connection-filter-postgis
-
-# [2.3.0](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter-postgis@2.2.4...graphile-plugin-connection-filter-postgis@2.3.0) (2026-03-01)
-
-**Note:** Version bump only for package graphile-plugin-connection-filter-postgis
-
-## [2.2.4](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter-postgis@2.2.3...graphile-plugin-connection-filter-postgis@2.2.4) (2026-02-28)
-
-**Note:** Version bump only for package graphile-plugin-connection-filter-postgis
-
-## [2.2.3](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter-postgis@2.2.2...graphile-plugin-connection-filter-postgis@2.2.3) (2026-02-28)
-
-**Note:** Version bump only for package graphile-plugin-connection-filter-postgis
-
-## [2.2.2](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter-postgis@2.2.1...graphile-plugin-connection-filter-postgis@2.2.2) (2026-02-26)
-
-**Note:** Version bump only for package graphile-plugin-connection-filter-postgis
-
-## [2.2.1](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter-postgis@2.2.0...graphile-plugin-connection-filter-postgis@2.2.1) (2026-02-25)
-
-### Bug Fixes
-
-- pin all Graphile ecosystem RC versions and sync inconsistencies ([8f553c1](https://github.com/constructive-io/constructive/commit/8f553c10d0bdc868bc1ac34136a6dfdab76f61b5))
-
-# [2.2.0](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter-postgis@2.1.1...graphile-plugin-connection-filter-postgis@2.2.0) (2026-02-24)
-
-### Features
-
-- **graphile-postgis:** add v5 postgis plugin packages ([daa2f7c](https://github.com/constructive-io/constructive/commit/daa2f7caca22b37a38a7a34b99ad0f58a2c44999))
-
-# [2.1.0](https://github.com/constructive-io/constructive/compare/graphile-plugin-connection-filter-postgis@2.1.1...graphile-plugin-connection-filter-postgis@2.1.0) (2026-02-24)
-
-### Features
-
-- **graphile-postgis:** add v5 postgis plugin packages ([daa2f7c](https://github.com/constructive-io/constructive/commit/daa2f7caca22b37a38a7a34b99ad0f58a2c44999))
