--- graphile/graphile-plugin-connection-filter-postgis/package.json
@@ -1,62 +1,0 @@
-{
-  "name": "graphile-plugin-connection-filter-postgis",
-  "version": "2.5.2",
-  "author": "Constructive <developers@constructive.io>",
-  "description": "PostGIS operators for postgraphile-plugin-connection-filter (PostGraphile v5)",
-  "main": "index.js",
-  "module": "esm/index.js",
-  "types": "index.d.ts",
-  "homepage": "https://github.com/constructive-io/constructive",
-  "license": "MIT",
-  "publishConfig": {
-    "access": "public",
-    "directory": "dist"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/constructive-io/constructive"
-  },
-  "bugs": {
-    "url": "https://github.com/constructive-io/constructive/issues"
-  },
-  "scripts": {
-    "clean": "makage clean",
-    "prepack": "npm run build",
-    "build": "makage build",
-    "build:dev": "makage build --dev",
-    "lint": "eslint . --fix",
-    "test": "jest",
-    "test:watch": "jest --watch"
-  },
-  "peerDependencies": {
-    "graphile-build": "5.0.0-rc.4",
-    "graphile-config": "1.0.0-rc.5",
-    "graphile-postgis": "^2.0.0",
-    "graphql": "^16.9.0",
-    "pg-sql2": "5.0.0-rc.4",
-    "postgraphile": "5.0.0-rc.7",
-    "graphile-connection-filter": "workspace:^"
-  },
-  "peerDependenciesMeta": {
-    "graphile-postgis": {
-      "optional": false
-    }
-  },
-  "devDependencies": {
-    "@types/node": "^22.19.11",
-    "graphile-postgis": "workspace:^",
-    "graphile-test": "workspace:^",
-    "makage": "^0.1.10"
-  },
-  "keywords": [
-    "postgraphile",
-    "graphql",
-    "postgresql",
-    "postgis",
-    "connection-filter",
-    "spatial",
-    "geometry",
-    "geography",
-    "constructive"
-  ]
-}
