--- graphile/graphile-plugin-connection-filter-postgis/tsconfig.esm.json
@@ -1,7 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "dist/esm",
-    "module": "ESNext"
-  }
-}
