+++ graphile/graphile-postgis/package.json
@@ -46,9 +46,15 @@
     "graphile-build": "5.0.0-rc.4",
     "graphile-build-pg": "5.0.0-rc.5",
     "graphile-config": "1.0.0-rc.5",
+    "graphile-connection-filter": "workspace:^",
     "graphql": "^16.9.0",
     "pg-sql2": "5.0.0-rc.4",
     "postgraphile": "5.0.0-rc.7"
+  },
+  "peerDependenciesMeta": {
+    "graphile-connection-filter": {
+      "optional": true
+    }
   },
   "devDependencies": {
     "@types/geojson": "^7946.0.14",
